+++ core/src/main/java/com/ensody/reactivestate/StateFlowStore.kt
@@ -28,6 +28,3 @@ public class InMemoryStateFlowStore : StateFlowStore {
             data
         } as MutableValueFlow<T>
 }
-
-/** A factory function creating a StateFlowStore. */
-public typealias StateFlowStoreFactory = (CoroutineScope) -> StateFlowStore