+++ reactivestate/src/test/java/com/ensody/reactivestate/android/LiveDataObserverTest.kt
@@ -1,9 +1,10 @@
-package com.ensody.reactivestate
+package com.ensody.reactivestate.android
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
 import assertk.assertThat
 import assertk.assertions.isEqualTo
+import com.ensody.reactivestate.autoRun
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runBlockingTest